move move request tests to scenarios, refactored
